requirements.txt
Flask==2.0.3
neo4j==4.3.6
openai==1.30.1
python-dotenv==0.19.2
gunicorn==20.1.0
Werkzeug==2.0.3
tenacity==8.0.1
termcolor==1.1.0